Successfully nested Exercise RecyclerView inside Day RecyclerView
For the WorkoutDay UI, a list of exercise must be available inside a list of days. Therefore a RecyclerView was needed to be nested inside another RecyclerView. This was accomplished by creating the Exercise Adapter and ViewHolder inside the WorkoutDay ViewHolder. The effect of this is that now there is a horizontal scrolling list similar to how apps appear in the Google Play Main screen.

It was expected that an issue with the RecyclerView not being properly inflated would be caused, see: https://stackoverflow.com/questions/34569217/how-to-add-a-recyclerview-inside-another-recyclerview

However this was avoided, presumably by hardcoding the width of each ExerciseDetail View. This was required anyway.