Launch workout feature added
A basic "launch workout" feature has been added. That is to say, the basics are. The feature doesn't actually save data yet as that entire process will be implemented later. However the basic prototype of how it's supposed to work has been added. Only a few fixes here and there remain, not counting the database integeration that is. The structure at this time of the data is messy at best and needs some serious refactoring which will come later.

This current branch focuses only on the "start workout" feature. It does not include changes to any other features like "add / edit workout", etc.