Made some major changes with the UI and Dialogs
This commit brings several updates. The first being that Exercise Comments are now wired up to the rest of the Activities and so is the Exercise Guide feature. These activities are not yet hooked up to Firebase however. This comment also brings changes to the default Button style which has now been updated for all Activities. Structural changes are also made in the way the code uses DialogFragments. There is now a common, custom built EditDialogFragment that receives a callback from the Activity so that when the Dialog completes, the Activity can handle it from there.