Refactored ExerciseList stuff as much as I could
Before this commit a boolean was passed to the ExerciseListActivity which then passed it to their fragments. These Fragments would show a switch or an edit/delete menu depending on where it was launched from. Now there are two ListActivity subclasses which are provided different ViewHolders from a Factory method in the ExerciseListActivity superclass. This approach is somewhat more polymorphic. However, an enum still had to be used in the method because they can be easily serialized.