Split EditExercise into a couple of classes
This commit splits the EditExerciseActivity into a superclass and a subclass for adding a new Activity. It was intended that there be two Activities for Lifting and Cardio exercises but at this point that does seem unecessary. Primarily because the type of exercise is determined by a Radio button and it doesn't seem like a good idea to load a different activity when a RadioButton is clicked. However, this may change in the future.