Work has begun on integrating ActiveWorkout
This commit starts the process of Firebase integration in ActiveWorkout. A few changes to certain models is made along with proper integration with Firebase. However, as a whole, the feature now takes data from Firebase. Work still remains though. As of this commit, there is no provision of getting targets from a "previous date" of when the exercise was performed. That being said, the MockData layer is slowly being phased out and proper data is being put in its place.