Added a proper fragment at the end of workouts
This commit makes adds a fragment to the end of a workout. This fragment will be the one responsible to save all the points data, when the user finally terminates thier workout. We cannot save the points data before that otherwise, every time the user saves, the points will be added to all the proper nodes.