Exercise guide images can be added
This commit adds the functionality to add guide images to exercises. However the interface to view and get to them is terrible. The UI will be changed to a more sane, usable version at a later date. Working with the URIs and getting a path needed a UriPathUtil class based on the FileUtils class for URIs that is quite popular on github. to be created which has code to get the path from various URI schemes and such. At this point I do not understand it fully but this was the only way that worked. This class has a lot of uneccessary code I believe and needs to be slimmed down but I could be wrong about that.